Added resource limits to deployment strategy
So that the deployer pods are limited to 256mi (instead of 512mi default).
Interestingly on OSO its not possible to have a limit less than 256mi,
otherwise the deployment stuck as 'New' and do not proceed
(even not after an hour).

As bonus 'syndesis-datamapper' file has been renamed to 'syndesis-atlasamap'.